+++ controllers/tasks.js
@@ -3,8 +3,9 @@ const taskModel = require('../models/tasksModel')
 // let todoTasks = [] // local tasks
 
 exports.getTasks = async (req, res) => {
+  const { id } = req.usuario
   try {
-    const tasks = await taskModel.find()
+    const tasks = await taskModel.find({ creador: id })
     if (tasks.length === 0) {
       return res.json({ msg: 'Todo task empty' })
     }
@@ -15,6 +16,7 @@ exports.getTasks = async (req, res) => {
 }
 
 exports.postTasks = (req, res) => {
+  const { id } = req.usuario
   try {
     if (Object.keys(req.body).length === 0) {
       return res.json({ msg: 'The task is empty, please fill the fields :)' })
@@ -22,6 +24,7 @@ exports.postTasks = (req, res) => {
     // eslint-disable-next-line new-cap
     const task = new taskModel(req.body)
     // task.id = uuid4()
+    task.creador = id
     task.save()
     res.json({ msg: 'Added Successful' })
   } catch (error) {
@@ -36,6 +39,9 @@ exports.deleteTasks = async (req, res) => {
     if (!taskEliminated) {
       return res.status(400).json({ msg: 'There is any task to eliminate' })
     }
+    if (taskEliminated.creador !== req.usuario.id) {
+      return res.status(400).json({ msg: 'You can\'t eliminate this task' })
+    }
     // todoTasks = todoTasks.filter(item => item.id !== id) // local task
     await taskModel.findByIdAndDelete({ _id: id })
     res.json({ msg: 'Task was deleted' })
@@ -46,7 +52,8 @@ exports.deleteTasks = async (req, res) => {
 
 exports.deleteAllTasks = async (req, res) => {
   try {
-    await taskModel.deleteMany()
+    const { id } = req.usuario
+    await taskModel.deleteMany({ creador: id })
     res.json({ msg: 'All tasks was deleted' })
   } catch (error) {
     console.log(error)
@@ -59,6 +66,9 @@ exports.updateTasks = async (req, res) => {
     const taskUpdated = await taskModel.findById(id)
     if (!taskUpdated) {
       return res.json(400).json({ msg: 'This task does not exist' })
+    }
+    if (taskUpdated.creador !== req.usuario.id) {
+      return res.status(400).json({ msg: 'You can\'t eliminate this task' })
     }
     const { title, description, finished } = req.body
     // eslint-disable-next-line no-return-assign
